chore(src): reorganize Search elements
