+++ src/helix-ui/elements/hx-search/_shadow.less
@@ -1,5 +1,5 @@
 @import (reference) "mixins";
-@import "./HXElement.less";
+@import "../HXElement";
 
 // TODO: add CSS custom props for theming hooks
