+++ src/helix-ui/elements/hx-search/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXSearchElement.html';
-import shadowStyles from './HXSearchElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 /**
  * Fires when the element loses focus.
